Refactor the product action buttons to remove useService in userland
