+++ src/components/store/ProductDetails.tsx
@@ -4,6 +4,7 @@ import {
   ProductModifiers,
   ProductVariantSelector,
   SelectedVariant,
+  ProductActions,
 } from '../../headless/store/components';
 import { SelectedVariantServiceDefinition } from '../../headless/store/services/selected-variant-service';
 import { ProductActionButtons } from './ProductActionButtons';
@@ -631,11 +632,8 @@ export default function ProductDetails({
 
           {/* Add to Cart */}
           <div className="space-y-4">
-            <ProductVariantSelector.Trigger>
+            <ProductActions.Actions>
               {({
-                onAddToCart,
-                canAddToCart,
-                isLoading,
                 error,
                 isPreOrderEnabled,
                 preOrderMessage,
@@ -656,17 +654,12 @@ export default function ProductDetails({
                   )}
 
                   <ProductActionButtons
-                    onAddToCart={onAddToCart}
-                    canAddToCart={canAddToCart}
-                    isLoading={isLoading}
-                    isPreOrderEnabled={isPreOrderEnabled}
-                    inStock={inStock}
                     onShowSuccessMessage={setShowSuccessMessage}
                     isQuickView={isQuickView}
                   />
                 </div>
               )}
-            </ProductVariantSelector.Trigger>
+            </ProductActions.Actions>
 
             {/* Stock Status */}
             <ProductVariantSelector.Stock>
