+++ src/headless/store/components/ProductVariantSelector.tsx
@@ -180,106 +180,6 @@ export const Choice = (props: ChoiceProps) => {
 };
 
 /**
- * Props for Trigger headless component
- */
-export interface TriggerProps {
-  /** Quantity to add (optional) */
-  quantity?: number;
-  /** Render prop function that receives trigger data */
-  children: (props: TriggerRenderProps) => React.ReactNode;
-}
-
-/**
- * Render props for Trigger component
- */
-export interface TriggerRenderProps {
-  /** Function to add product to cart */
-  onAddToCart: () => Promise<void>;
-  /** Whether add to cart is available */
-  canAddToCart: boolean;
-  /** Whether add to cart is currently loading */
-  isLoading: boolean;
-  /** Current variant price */
-  price: string;
-  /** Whether variant is in stock */
-  inStock: boolean;
-  /** Whether pre-order is enabled */
-  isPreOrderEnabled: boolean;
-  /** Pre-order message */
-  preOrderMessage: string | null;
-  /** Error message if any */
-  error: string | null;
-  /** Available quantity */
-  availableQuantity: number | null;
-}
-
-/**
- * Headless component for add to cart trigger
- */
-export const Trigger = (props: TriggerProps) => {
-  const variantService = useService(
-    SelectedVariantServiceDefinition
-  ) as ServiceAPI<typeof SelectedVariantServiceDefinition>;
-
-  // Try to get modifiers service - it may not exist for all products
-  let modifiersService: ServiceAPI<
-    typeof ProductModifiersServiceDefinition
-  > | null = null;
-  try {
-    modifiersService = useService(
-      ProductModifiersServiceDefinition
-    ) as ServiceAPI<typeof ProductModifiersServiceDefinition>;
-  } catch {
-    // Modifiers service not available for this product
-    modifiersService = null;
-  }
-
-  const price = variantService.currentPrice.get();
-  const inStock = variantService.isInStock.get();
-  const isPreOrderEnabled = variantService.isPreOrderEnabled.get();
-  const preOrderMessage = variantService.preOrderMessage.get();
-  const isLoading = variantService.isLoading.get();
-  const error = variantService.error.get();
-  const availableQuantity = variantService.quantityAvailable.get();
-  const quantity = variantService.selectedQuantity.get();
-
-  // Check if all required modifiers are filled
-  const areAllRequiredModifiersFilled = modifiersService
-    ? modifiersService.areAllRequiredModifiersFilled()
-    : true; // If no modifiers service, assume no required modifiers
-
-  const canAddToCart =
-    (inStock || isPreOrderEnabled) &&
-    !isLoading &&
-    areAllRequiredModifiersFilled;
-
-  const onAddToCart = async () => {
-    // Get modifiers data if available
-    let modifiersData: Record<string, any> | undefined;
-    if (modifiersService) {
-      const selectedModifiers = modifiersService.selectedModifiers.get();
-      if (Object.keys(selectedModifiers).length > 0) {
-        modifiersData = selectedModifiers;
-      }
-    }
-
-    await variantService.addToCart(quantity, modifiersData);
-  };
-
-  return props.children({
-    onAddToCart,
-    canAddToCart,
-    isLoading,
-    price,
-    inStock,
-    isPreOrderEnabled,
-    preOrderMessage,
-    error,
-    availableQuantity,
-  });
-};
-
-/**
  * Props for Stock headless component
  */
 export interface StockProps {
